CRE-482 - Sets up infra on normal funcs
